requirements.txt
fastapi
uvicorn
sqlalchemy

alembic
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
